+++ lint.bash
@@ -6,7 +6,6 @@ shopt -s globstar
 
 targets=(src/**/*.[ch]pp test/**/*.[ch]pp)
 
-cppcheck --language=c++ "${targets[@]}" || exit 1
-cpplint --linelength=120 --filter=-legal/copyright,-build/c++11 "${targets[@]}" || exit 1
+cpplint --linelength=120 --filter=-build/include_subdir,-legal/copyright,-build/c++11 "${targets[@]}" || exit 1
 misspell -error "${targets[@]}" || exit 1
 shellcheck ./**/*.bash || exit 1