+++ src/muxer/fdk_aac_audio_producer.hpp
@@ -1,0 +1,19 @@
+#pragma once
+
+#include "muxer/audio_producer.hpp"
+
+namespace hisui {
+
+class Config;
+class Metadata;
+
+}  // namespace hisui
+
+namespace hisui::muxer {
+
+class FDKAACAudioProducer : public AudioProducer {
+ public:
+  FDKAACAudioProducer(const hisui::Config&, const hisui::Metadata& i);
+};
+
+}  // namespace hisui::muxer