+++ src/video/buffer_vpx_encoder.hpp
@@ -7,10 +7,17 @@
 #include <queue>
 #include <vector>
 
+#include <boost/cstdint.hpp>
 #include <boost/rational.hpp>
 
+#include "constants.hpp"
 #include "video/encoder.hpp"
-#include "webm/output/context.hpp"
+
+namespace hisui {
+
+struct Frame;
+
+}
 
 namespace hisui::video {
 
@@ -18,8 +25,10 @@ class VPXEncoderConfig;
 
 class BufferVPXEncoder : public Encoder {
  public:
-  BufferVPXEncoder(std::queue<hisui::webm::output::FrameTuple>*,
-                   const VPXEncoderConfig& config);
+  BufferVPXEncoder(
+      std::queue<hisui::Frame>*,
+      const VPXEncoderConfig&,
+      const std::uint64_t timescale = hisui::Constants::NANO_SECOND);
   void outputImage(const std::vector<unsigned char>&);
   void flush();
   ~BufferVPXEncoder();
@@ -27,7 +36,7 @@ class BufferVPXEncoder : public Encoder {
   std::uint32_t getFourcc() const;
 
  private:
-  std::queue<hisui::webm::output::FrameTuple>* m_buffer;
+  std::queue<hisui::Frame>* m_buffer;
   std::uint32_t m_width;
   std::uint32_t m_height;
   boost::rational<std::uint64_t> m_fps;
@@ -36,6 +45,7 @@ class BufferVPXEncoder : public Encoder {
   ::vpx_codec_ctx_t m_codec;
   ::vpx_image_t m_raw_vpx_image;
   std::uint64_t m_sum_of_bits = 0;
+  const std::uint64_t m_timescale;
 
   bool encodeFrame(::vpx_codec_ctx_t*, ::vpx_image_t*, const int, const int);
 };